+++ modules/apps/commerce/commerce-inventory-api/src/main/java/com/liferay/commerce/inventory/service/CommerceInventoryWarehouseItemService.java
@@ -16,6 +16,8 @@ import com.liferay.portal.kernel.transaction.Isolation;
 import com.liferay.portal.kernel.transaction.Propagation;
 import com.liferay.portal.kernel.transaction.Transactional;
 
+import java.math.BigDecimal;
+
 import java.util.Date;
 import java.util.List;
 
@@ -46,13 +48,13 @@ public interface CommerceInventoryWarehouseItemService extends BaseService {
 	 */
 	public CommerceInventoryWarehouseItem addCommerceInventoryWarehouseItem(
 			String externalReferenceCode, long commerceInventoryWarehouseId,
-			int quantity, String sku, String unitOfMeasureKey)
+			BigDecimal quantity, String sku, String unitOfMeasureKey)
 		throws PortalException;
 
 	public CommerceInventoryWarehouseItem
 			addOrUpdateCommerceInventoryWarehouseItem(
 				String externalReferenceCode, long companyId,
-				long commerceInventoryWarehouseId, int quantity, String sku,
+				long commerceInventoryWarehouseId, BigDecimal quantity, String sku,
 				String unitOfMeasureKey)
 		throws PortalException;
 
@@ -143,28 +145,29 @@ public interface CommerceInventoryWarehouseItemService extends BaseService {
 	public String getOSGiServiceIdentifier();
 
 	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
-	public int getStockQuantity(long companyId, long groupId, String sku);
+	public BigDecimal getStockQuantity(
+		long companyId, long groupId, String sku);
 
 	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
-	public int getStockQuantity(long companyId, String sku);
+	public BigDecimal getStockQuantity(long companyId, String sku);
 
 	public CommerceInventoryWarehouseItem
 			increaseCommerceInventoryWarehouseItemQuantity(
-				long commerceInventoryWarehouseItemId, int quantity)
+				long commerceInventoryWarehouseItemId, BigDecimal quantity)
 		throws PortalException;
 
 	public void moveQuantitiesBetweenWarehouses(
 			long fromCommerceInventoryWarehouseId,
-			long toCommerceInventoryWarehouseId, int quantity, String sku)
+			long toCommerceInventoryWarehouseId, BigDecimal quantity, String sku)
 		throws PortalException;
 
 	public CommerceInventoryWarehouseItem updateCommerceInventoryWarehouseItem(
-			long commerceInventoryWarehouseItemId, int quantity,
-			int reservedQuantity, long mvccVersion)
+			long commerceInventoryWarehouseItemId, BigDecimal quantity,
+			BigDecimal reservedQuantity, long mvccVersion)
 		throws PortalException;
 
 	public CommerceInventoryWarehouseItem updateCommerceInventoryWarehouseItem(
-			long commerceInventoryWarehouseItemId, int quantity,
+			long commerceInventoryWarehouseItemId, BigDecimal quantity,
 			long mvccVersion)
 		throws PortalException;
